refactor: extract usePullToRefresh hook, remove dead code, fix SK security hotspot
