+++ app/src/screens/MyEventsScreen.js
@@ -1,7 +1,7 @@
 import { Ionicons } from '@expo/vector-icons';
 import { FlashList } from '@shopify/flash-list';
 import { collection, deleteDoc, doc, onSnapshot, query, where } from 'firebase/firestore';
-import React, { useEffect, useState, useCallback, useRef } from 'react';
+import React, { useEffect, useState, useCallback } from 'react';
 import {
     ActivityIndicator,
     Alert,
@@ -14,6 +14,7 @@ import {
 import EventCard from '../components/EventCard';
 import LiquidPullToRefresh from '../components/LiquidPullToRefresh';
 import ScreenWrapper from '../components/ScreenWrapper';
+import usePullToRefresh from '../hooks/usePullToRefresh';
 import { useAuth } from '../lib/AuthContext';
 import { useTheme } from '../lib/ThemeContext';
 import { db } from '../lib/firebaseConfig';
@@ -26,21 +27,10 @@ export default function MyEventsScreen({ navigation }) {
     const [loading, setLoading] = useState(true);
     const [refreshing, setRefreshing] = useState(false);
     const [refreshNonce, setRefreshNonce] = useState(0);
-    const [pullDistance, setPullDistance] = useState(0);
-    const lastPullRef = useRef(0);
-
-    const handleScroll = useCallback(e => {
-        const offsetY = e.nativeEvent.contentOffset.y;
-        lastPullRef.current = Math.max(0, -offsetY);
-        setPullDistance(lastPullRef.current);
-    }, []);
-
-    const handleScrollEndDrag = useCallback(() => {
-        if (lastPullRef.current >= 80 && !refreshing) {
-            setRefreshing(true);
-            setRefreshNonce(n => n + 1);
-        }
-    }, [refreshing]);
+    const { pullDistance, handleScroll, handleScrollEndDrag } = usePullToRefresh(refreshing, () => {
+        setRefreshing(true);
+        setRefreshNonce(n => n + 1);
+    });
 
     useEffect(() => {
         if (!user) return;
@@ -95,11 +85,6 @@ export default function MyEventsScreen({ navigation }) {
                 },
             ]);
         }
-    };
-
-    const onRefresh = () => {
-        setRefreshing(true);
-        setRefreshNonce(n => n + 1);
     };
 
     // 🚀 Task 3: Wrap component renderer with useCallback to avoid functional rebuilds on updates
