+++ app/src/screens/MyRegisteredEventsScreen.js
@@ -1,9 +1,10 @@
 import { Ionicons } from '@expo/vector-icons';
 import { collection, documentId, getDocs, onSnapshot, query, where } from 'firebase/firestore';
-import { useCallback, useEffect, useRef, useState } from 'react';
+import { useEffect, useState } from 'react';
 import { ActivityIndicator, FlatList, StyleSheet, Text, View } from 'react-native';
 import EventCard from '../components/EventCard';
 import LiquidPullToRefresh from '../components/LiquidPullToRefresh';
+import usePullToRefresh from '../hooks/usePullToRefresh';
 import { useAuth } from '../lib/AuthContext';
 import { db } from '../lib/firebaseConfig';
 import { useTheme } from '../lib/ThemeContext';
@@ -16,21 +17,10 @@ export default function MyRegisteredEventsScreen() {
     const [loading, setLoading] = useState(true);
     const [refreshing, setRefreshing] = useState(false);
     const [refreshNonce, setRefreshNonce] = useState(0);
-    const [pullDistance, setPullDistance] = useState(0);
-    const lastPullRef = useRef(0);
-
-    const handleScroll = useCallback(e => {
-        const offsetY = e.nativeEvent.contentOffset.y;
-        lastPullRef.current = Math.max(0, -offsetY);
-        setPullDistance(lastPullRef.current);
-    }, []);
-
-    const handleScrollEndDrag = useCallback(() => {
-        if (lastPullRef.current >= 80 && !refreshing) {
-            setRefreshing(true);
-            setRefreshNonce(n => n + 1);
-        }
-    }, [refreshing]);
+    const { pullDistance, handleScroll, handleScrollEndDrag } = usePullToRefresh(refreshing, () => {
+        setRefreshing(true);
+        setRefreshNonce(n => n + 1);
+    });
 
     useEffect(() => {
         if (!user) {
@@ -89,11 +79,6 @@ export default function MyRegisteredEventsScreen() {
 
         return () => unsubscribe();
     }, [user, refreshNonce]);
-
-    const onRefresh = () => {
-        setRefreshing(true);
-        setRefreshNonce(n => n + 1);
-    };
 
     if (loading) {
         return (
